requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
jsobject==0.10.2
MarkupSafe==1.1.0
mysql-connector-python==8.0.13
protobuf==3.6.1
six==1.11.0
Werkzeug==0.14.1
wheel==0.24.0
